refactor: move just-bash agent environment to environment-registry (#118)
Move JustBashAgentEnvironment implementation, tools, utilities, and tests
from @otter-agent/core to @otter-agent/environment-registry.

- Move all just-bash source files to packages/environment-registry/src/environments/just-bash/
- Create barrel file (index.ts) exporting only JustBashAgentEnvironment, JustBashAgentEnvironmentTemplate, and public types
- Register just-bash in the environment registry (register-environments.ts)
- Rename packages/otter-agent/src/environments/ to environment/ containing only createEnvironmentExtension
- Remove AgentEnvironment namespace (declaration merging) and isJustBashAgentEnvironment
- Duplicate escapeXml utility into both packages (escape-xml.ts)
- Move just-bash and diff dependencies from core to environment-registry
- Update JSDoc examples to use isSkillSupportedAgentEnvironment
