+++ packages/environment-registry/src/environments/just-bash/edit-diff.ts
