+++ packages/environment-registry/src/environments/just-bash/file-mutation-queue.ts
