+++ packages/environment-registry/src/environments/just-bash/index.ts
@@ -1,0 +1,8 @@
+export {
+	JustBashAgentEnvironment,
+	JustBashAgentEnvironmentTemplate,
+} from "./just-bash-agent-environment.js";
+export type {
+	JustBashAgentEnvironmentOptions,
+	JustBashToolName,
+} from "./just-bash-agent-environment.js";
