+++ packages/environment-registry/src/environments/just-bash/path-utils.ts
