+++ packages/environment-registry/src/environments/just-bash/truncate.ts
