+++ packages/environment-registry/src/utils/escape-xml.ts
