--- packages/otter-agent/src/environments/index.ts
@@ -1,10 +1,0 @@
-export {
-	AgentEnvironment,
-	JustBashAgentEnvironment,
-	JustBashAgentEnvironmentOptionsSchema,
-	JustBashAgentEnvironmentTemplate,
-	isJustBashAgentEnvironment,
-	type JustBashAgentEnvironmentOptions,
-	type JustBashToolName,
-} from "./agent-environment.js";
-export { createEnvironmentExtension } from "./environment-extension.js";
