+++ packages/otter-agent/src/utils/escape-xml.ts
@@ -1,0 +1,9 @@
+/** Escape characters that are special in XML/HTML attributes and text content. */
+export function escapeXml(str: string): string {
+	return str
+		.replace(/&/g, "&amp;")
+		.replace(/</g, "&lt;")
+		.replace(/>/g, "&gt;")
+		.replace(/"/g, "&quot;")
+		.replace(/'/g, "&apos;");
+}
